bin component filter added so you dont have to reload the page all the time

+++ client/src/app/bin/bin.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
-import {UserService} from '../../_service/user.service';
-import {NgFor, NgIf} from '@angular/common';
-import {MatIcon} from '@angular/material/icon';
+import { UserService } from '../../_service/user.service';
+import { NgFor, NgIf } from '@angular/common';
+import { MatIcon } from '@angular/material/icon';
 
 @Component({
   selector: 'app-bin',
@@ -14,29 +14,50 @@ export class BinComponent {
   users: any = [];
 
   constructor(private userService: UserService) {
+    this.loadDeletedUsers();
+  }
+
+  loadDeletedUsers() {
     this.userService.getDeletedUsers().subscribe({
       next: (data) => {
         this.users = data;
-        console.log(this.users);
+      },
+      error: (err) => {
+        console.error("Error fetching deleted users:", err);
       }
-    })
-
+    });
   }
 
   restore(id: number) {
-    this.userService.restoreUser(id).subscribe({
-      next: (employeeData) => {
-        console.log(employeeData);
-      }
-    })
+    const confirmRestore = window.confirm("Möchten Sie diesen Benutzer wirklich wiederherstellen?");
+    if (confirmRestore) {
+      this.userService.restoreUser(id).subscribe({
+        next: () => {
+          this.users = this.users.filter((user: { id: number; }) => user.id !== id);
+          alert("Benutzer wurde erfolgreich wiederhergestellt!");
+        },
+        error: (err) => {
+          console.error("Error restoring user:", err);
+          alert("Fehler beim Wiederherstellen des Benutzers.");
+        }
+      });
+    }
   }
 
   deletePermanently(id: number) {
-    this.userService.deleteUser(id).subscribe({
-      next: (employeeData) => {
-        console.log(employeeData);
-      }
-    })
+    const confirmDelete = window.confirm("Sind Sie sicher, dass Sie diesen Benutzer endgültig löschen möchten?");
+    if (confirmDelete) {
+      this.userService.deleteUser(id).subscribe({
+        next: () => {
+          this.users = this.users.filter((user: { id: number; }) => user.id !== id);
+          alert("Benutzer wurde dauerhaft gelöscht!");
+        },
+        error: (err) => {
+          console.error("Error deleting user:", err);
+          alert("Fehler beim endgültigen Löschen des Benutzers.");
+        }
+      });
+    }
   }
 
   goBack() {
@@ -46,5 +67,4 @@ export class BinComponent {
   logout() {
     this.userService.logout();
   }
-
 }
